Cleaned up Sound class.
I changed the string parameters for SetFile, SetSet, and the constructor to 'const std::string&' so it is being sent as a reference instead. Before the changes it would make a copy of the string being sent as an argument. And the const basically marks it as not being changed in the scope of the function/constructor.